Move cb_with_observations_label to separate file
